refactor(*.*): 코멘트 반영, master 머지
