+++ src/main/java/com/deliverysaurus/config/SpringSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.jordyshop.config;
+package com.deliverysaurus.config;
 
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
