+++ src/main/java/com/deliverysaurus/member/domain/Address.java
@@ -3,10 +3,11 @@ package com.deliverysaurus.member.domain;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
+import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Embeddable
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Address {
 
     @Column(name = "address")
