+++ src/main/java/com/deliverysaurus/member/domain/Name.java
@@ -3,10 +3,11 @@ package com.deliverysaurus.member.domain;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
+import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Embeddable
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Name {
     private static final int NAME_MAX_LENGTH = 5;
