+++ src/test/java/com/deliverysaurus/config/EnumController.java
@@ -1,4 +1,4 @@
-package com.example.jordyshop.domain;
+package com.deliverysaurus.config;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.Arrays;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import com.deliverysaurus.member.domain.Gender;
 
 @RestController
 public class EnumController {
